refine logging and messages
Refine the syslog error messages to inform the user that an error
occurred and also send a copy of messages directed to the user standard
error stream to the system logger. This will make the journal entries
easier to follow.

Remove the error count on exit since it is not relevant now that
each error prints a message, and only print the error count in the
interactive prompt if at least one error occurred.

From now on, the message macros will be used as follow:

  - info() for informative messages in the system journal, such as program steps;
  - notice() for non-error highlights in the system journal;
  - crit() for system call failure, handling errno;
  - err() for non-critical errors, such as bad syntax;
  - debug() for function debugging purpose;
  - user() for writing to the user standard error stream;
  - print() is introduced for writing to standard output.

Even though a critical error means that the system is in bad shape, the
shell is still functional, remove the note about exiting in such case.