report invocation error to the parent
If the execution of a child process failed, wrap the error into a
standard return value so that the parent can log the error.

Similarly to a standard shell, use 127 for file not found, and 126
for permission errors but also general errors.

Note that a process can execute correctly but its subprocess might
return 127 for example. But still reporting no such file from our
shell is totally fine.